+++ src/main/java/com/infoshare/academy/highfive/consolemenu/HolidaysReviewMenu.java
@@ -1,27 +1,26 @@
-package com.infoshare.academy.highfive.menuconsola;
+package com.infoshare.academy.highfive.consolemenu;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.util.Scanner;
 
-public class ConfigurationMenu extends MainMenu {
+public class HolidaysReviewMenu extends MainMenu {
     private static final Logger stdout = LoggerFactory.getLogger("CONSOLE_OUT");
 
     public static void runSubmenu() throws Exception {
 
-        ConfigurationMenu configurationMenu = new ConfigurationMenu();
-        configurationMenu.menuOptionsDisplay();
-        configurationMenu.getUserChoice();
+        HolidaysReviewMenu holidaysReviewMenu = new HolidaysReviewMenu();
+        holidaysReviewMenu.menuOptionsDisplay();
+        holidaysReviewMenu.getUserChoice();
     }
 
     @Override
     public void menuOptionsDisplay() {
 
-        menuOptions.add("CONFIGURATION");
-        menuOptions.add("Change DATE format");
-        menuOptions.add("Change employees SORTING mode");
-        menuOptions.add("Change configuration of external file");
+        menuOptions.add("HOLIDAYS REVIEW");
+        menuOptions.add("Display employee vacation");
+        menuOptions.add("Display team vacations");
         menuOptions.add("Previous menu");
 
         stdout.info("\n\n" + menuOptions.get(0) + "\n\n");
@@ -30,7 +29,7 @@ public class ConfigurationMenu extends MainMenu {
 
             stdout.info(i + ": " + menuOptions.get(i) + "\n");
         }
-    }
+     }
 
     @Override
     public int getUserChoice() throws Exception {
@@ -53,13 +52,10 @@ public class ConfigurationMenu extends MainMenu {
         } else {
             switch (userChoice) {
                 case 1:
-                    stdout.info("\nChange DATE format - UNDER CONSTRUCTION\n\n");
+                    stdout.info("\nDisplay employee vacation - UNDER CONSTRUCTION\n\n");
                     break;
                 case 2:
-                    stdout.info("\nChange employees SORTING mode - UNDER CONSTRUCTION\n\n");
-                    break;
-                case 3:
-                    stdout.info("\nChange configuration of external file - UNDER CONSTRUCTION\n\n");
+                    stdout.info("\nDisplay team vacations - UNDER CONSTRUCTION\n\n");
                     break;
                 default:
                     MainMenu.runMenu();
