+++ src/main/java/com/infoshare/academy/highfive/consolemenu/MainMenu.java
@@ -1,27 +1,38 @@
-package com.infoshare.academy.highfive.menuconsola;
+package com.infoshare.academy.highfive.consolemenu;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Scanner;
 
-public class HolidaysReviewMenu extends MainMenu {
+public class MainMenu {
     private static final Logger stdout = LoggerFactory.getLogger("CONSOLE_OUT");
+    public static int userChoice;
+    List<String> menuOptions = new ArrayList<>();
+    public String userChoiceString;
 
-    public static void runSubmenu() throws Exception {
 
-        HolidaysReviewMenu holidaysReviewMenu = new HolidaysReviewMenu();
-        holidaysReviewMenu.menuOptionsDisplay();
-        holidaysReviewMenu.getUserChoice();
+    public static void runMenu() throws Exception {
+
+        MainMenu mainMenu = new MainMenu();
+
+        do {
+            mainMenu.menuOptionsDisplay();
+            userChoice = mainMenu.getUserChoice();
+        }
+        while (userChoice != 0);
     }
 
-    @Override
     public void menuOptionsDisplay() {
 
+        menuOptions.clear();
+        menuOptions.add("MAIN MENU");
+        menuOptions.add("HOLIDAY PLANNING");
         menuOptions.add("HOLIDAYS REVIEW");
-        menuOptions.add("Display employee vacation");
-        menuOptions.add("Display team vacations");
-        menuOptions.add("Previous menu");
+        menuOptions.add("EMPLOYEES MANAGING");
+        menuOptions.add("CONFIGURATION");
 
         stdout.info("\n\n" + menuOptions.get(0) + "\n\n");
 
@@ -29,9 +40,9 @@ public class HolidaysReviewMenu extends MainMenu {
 
             stdout.info(i + ": " + menuOptions.get(i) + "\n");
         }
-     }
+        stdout.info("\n" + "0: Exit" + "\n");
+    }
 
-    @Override
     public int getUserChoice() throws Exception {
 
         boolean matchedToPattern;
@@ -46,19 +57,26 @@ public class HolidaysReviewMenu extends MainMenu {
         }
         userChoice = Integer.parseInt(userChoiceString);
 
-        if (userChoice > menuOptions.size() - 1 || userChoice == 0) {
+        if (userChoice > menuOptions.size() - 1) {
             stdout.info("Wrong input - try again\n");
             getUserChoice();
         } else {
             switch (userChoice) {
+                case 0: System.exit(0);
                 case 1:
-                    stdout.info("\nDisplay employee vacation - UNDER CONSTRUCTION\n\n");
+                    HolidaysPlanningMenu.runSubmenu();
                     break;
                 case 2:
-                    stdout.info("\nDisplay team vacations - UNDER CONSTRUCTION\n\n");
+                    HolidaysReviewMenu.runSubmenu();
+                    break;
+                case 3:
+                    EmployeesManagingMenu.runSubmenu();
+                    break;
+                case 4:
+                    ConfigurationMenu.runSubmenu();
                     break;
                 default:
-                    MainMenu.runMenu();
+                    break;
             }
         }
         return userChoice;
